+++ changelog.md
@@ -1,6 +1,7 @@
 #### 1.2.4 (2021-04-15)
 
-- Stay logged in in dev environments (#585cddc)
+- cleanup (#eb38aac)
+- Stay logged in in dev environments (#acad660)
 
 #### 1.2.3 (2021-04-14)